+++ platform/android/MapboxGLAndroidSDK/src/main/java/com/mapbox/mapboxsdk/maps/TrackingSettings.java
@@ -101,7 +101,7 @@ public class TrackingSettings {
      *
      * @return True to indicate the tracking modes will be dismissed.
      * @deprecated use @link #isAllDismissTrackingOnGestureinstead
-    */
+     */
     @Deprecated
     public boolean isDismissTrackingOnGesture() {
         return dismissLocationTrackingOnGesture && dismissBearingTrackingOnGesture;
@@ -215,7 +215,7 @@ public class TrackingSettings {
     private void validateGesturesForBearingTrackingMode() {
         int myBearingTrackingMode = getMyBearingTrackingMode();
         if (!dismissBearingTrackingOnGesture) {
-            if (myBearingTrackingMode == MyBearingTracking.NONE) {
+            if (myBearingTrackingMode == MyBearingTracking.NONE || myLocationTrackingMode == MyLocationTracking.TRACKING_NONE) {
                 uiSettings.setRotateGesturesEnabled(true);
             } else {
                 uiSettings.setRotateGesturesEnabled(false);