+++ platform/android/MapboxGLAndroidSDKTestApp/src/main/java/com/mapbox/mapboxsdk/testapp/activity/userlocation/MyLocationDrawableActivity.java
@@ -1,9 +1,14 @@
 package com.mapbox.mapboxsdk.testapp.activity.userlocation;
 
+import android.Manifest;
+import android.content.pm.PackageManager;
 import android.graphics.Color;
 import android.location.Location;
 import android.os.Bundle;
+import android.support.annotation.NonNull;
 import android.support.annotation.Nullable;
+import android.support.annotation.UiThread;
+import android.support.v4.app.ActivityCompat;
 import android.support.v4.content.ContextCompat;
 import android.support.v7.app.ActionBar;
 import android.support.v7.app.AppCompatActivity;
@@ -26,6 +31,8 @@ import com.mapbox.mapboxsdk.testapp.R;
 
 public class MyLocationDrawableActivity extends AppCompatActivity implements LocationListener {
 
+    private static final int PERMISSIONS_LOCATION = 0;
+
     private MapView mapView;
     private MapboxMap mapboxMap;
     private Location location;
@@ -45,20 +52,13 @@ public class MyLocationDrawableActivity extends AppCompatActivity implements Loc
         }
 
         findViewById(R.id.progress).setVisibility(View.GONE);
-        location = LocationServices.getLocationServices(this).getLastLocation();
 
         MapboxMapOptions mapboxMapOptions = new MapboxMapOptions();
         mapboxMapOptions.accessToken(getString(R.string.mapbox_access_token));
         mapboxMapOptions.styleUrl(Style.MAPBOX_STREETS);
-        mapboxMapOptions.locationEnabled(true);
-        mapboxMapOptions.camera(new CameraPosition.Builder()
-                .target(location != null ? new LatLng(location) : new LatLng(0, 0))
-                .zoom(11)
-                .tilt(25)
-                .build());
 
-        mapboxMapOptions.myLocationForegroundDrawables(ContextCompat.getDrawable(this, R.drawable.ic_chelsea),
-                ContextCompat.getDrawable(this, R.drawable.ic_chelsea));
+        // configure MyLocationView drawables
+        mapboxMapOptions.myLocationForegroundDrawable(ContextCompat.getDrawable(this, R.drawable.ic_chelsea));
         mapboxMapOptions.myLocationBackgroundDrawable(ContextCompat.getDrawable(this, R.drawable.ic_arsenal));
         mapboxMapOptions.myLocationForegroundTintColor(Color.GREEN);
         mapboxMapOptions.myLocationBackgroundTintColor(Color.YELLOW);
@@ -78,17 +78,53 @@ public class MyLocationDrawableActivity extends AppCompatActivity implements Loc
             @Override
             public void onMapReady(MapboxMap map) {
                 mapboxMap = map;
+                toggleGps(true);
             }
         });
+    }
 
-        LocationServices.getLocationServices(this).addLocationListener(this);
+    public void toggleGps(boolean enableGps) {
+        if (enableGps) {
+            if ((ContextCompat.checkSelfPermission(this, Manifest.permission.ACCESS_COARSE_LOCATION) != PackageManager.PERMISSION_GRANTED) ||
+                    (ContextCompat.checkSelfPermission(this, Manifest.permission.ACCESS_FINE_LOCATION) != PackageManager.PERMISSION_GRANTED)) {
+                ActivityCompat.requestPermissions(this, new String[]{Manifest.permission.ACCESS_COARSE_LOCATION, Manifest.permission.ACCESS_FINE_LOCATION}, PERMISSIONS_LOCATION);
+            } else {
+                enableLocation(true);
+            }
+        } else {
+            enableLocation(false);
+        }
+    }
+
+    private void enableLocation(boolean enabled) {
+        if (enabled) {
+            mapboxMap.setMyLocationEnabled(true);
+            Location location = mapboxMap.getMyLocation();
+            if (location != null) {
+                onLocationChanged(location);
+            }else{
+                LocationServices.getLocationServices(this).addLocationListener(this);
+            }
+        } else {
+            mapboxMap.setMyLocationEnabled(false);
+        }
+    }
+
+    @Override
+    public void onRequestPermissionsResult(int requestCode, @NonNull String permissions[], @NonNull int[] grantResults) {
+        switch (requestCode) {
+            case PERMISSIONS_LOCATION: {
+                if (grantResults.length > 0 && grantResults[0] == PackageManager.PERMISSION_GRANTED) {
+                    enableLocation(true);
+                }
+            }
+        }
     }
 
     @Override
     public void onLocationChanged(Location location) {
-        if (mapboxMap != null && firstRun) {
-            mapboxMap.moveCamera(CameraUpdateFactory.newLatLngZoom(new LatLng(location), 10));
-            firstRun = false;
+        if (mapboxMap != null) {
+            mapboxMap.moveCamera(CameraUpdateFactory.newLatLngZoom(new LatLng(location), 14));
         }
     }