+++ platform/android/MapboxGLAndroidSDKTestApp/src/main/java/com/mapbox/mapboxsdk/testapp/model/annotations/PulseMarkerViewOptions.java
@@ -7,27 +7,21 @@ import android.os.Parcelable;
 import com.mapbox.mapboxsdk.annotations.BaseMarkerViewOptions;
 import com.mapbox.mapboxsdk.annotations.Icon;
 import com.mapbox.mapboxsdk.annotations.IconFactory;
-import com.mapbox.mapboxsdk.annotations.MarkerView;
 import com.mapbox.mapboxsdk.geometry.LatLng;
 
-public class CountryMarkerViewOptions extends BaseMarkerViewOptions<CountryMarkerView, CountryMarkerViewOptions> {
+public class PulseMarkerViewOptions extends BaseMarkerViewOptions<PulseMarkerView, PulseMarkerViewOptions> {
 
-    private String abbrevName;
-    private int flagRes;
-
-    public CountryMarkerViewOptions() {
+    public PulseMarkerViewOptions() {
     }
 
-    protected CountryMarkerViewOptions(Parcel in) {
+    protected PulseMarkerViewOptions(Parcel in) {
         position((LatLng) in.readParcelable(LatLng.class.getClassLoader()));
         snippet(in.readString());
         title(in.readString());
         flat(in.readByte() != 0);
         anchor(in.readFloat(), in.readFloat());
-        infoWindowAnchor(in.readFloat(), in.readFloat());
+        selected = in.readByte() != 0;
         rotation(in.readFloat());
-        visible(in.readByte() != 0);
-        alpha(in.readFloat());
         if (in.readByte() != 0) {
             // this means we have an icon
             String iconId = in.readString();
@@ -35,12 +29,10 @@ public class CountryMarkerViewOptions extends BaseMarkerViewOptions<CountryMarke
             Icon icon = IconFactory.recreate(iconId, iconBitmap);
             icon(icon);
         }
-        abbrevName(in.readString());
-        flagRes(in.readInt());
     }
 
     @Override
-    public CountryMarkerViewOptions getThis() {
+    public PulseMarkerViewOptions getThis() {
         return this;
     }
 
@@ -59,32 +51,19 @@ public class CountryMarkerViewOptions extends BaseMarkerViewOptions<CountryMarke
         out.writeFloat(getAnchorV());
         out.writeFloat(getInfoWindowAnchorU());
         out.writeFloat(getInfoWindowAnchorV());
+        out.writeByte((byte) (selected ? 1 : 0));
         out.writeFloat(getRotation());
-        out.writeByte((byte) (isVisible() ? 1 : 0));
-        out.writeFloat(getAlpha());
         Icon icon = getIcon();
         out.writeByte((byte) (icon != null ? 1 : 0));
         if (icon != null) {
             out.writeString(getIcon().getId());
             out.writeParcelable(getIcon().getBitmap(), flags);
         }
-        out.writeString(abbrevName);
-        out.writeInt(flagRes);
     }
 
     @Override
-    public CountryMarkerView getMarker() {
-        return new CountryMarkerView(this, abbrevName, flagRes);
-    }
-
-    public CountryMarkerViewOptions abbrevName(String abbrevName) {
-        this.abbrevName = abbrevName;
-        return getThis();
-    }
-
-    public CountryMarkerViewOptions flagRes(int flagRes) {
-        this.flagRes = flagRes;
-        return getThis();
+    public PulseMarkerView getMarker() {
+        return new PulseMarkerView(this);
     }
 
     public static final Parcelable.Creator<CountryMarkerViewOptions> CREATOR
@@ -97,6 +76,4 @@ public class CountryMarkerViewOptions extends BaseMarkerViewOptions<CountryMarke
             return new CountryMarkerViewOptions[size];
         }
     };
-
-
 }